requirements.txt
streamlit
requests
Pillow
matplotlib
seaborn
numpy
gdown
scikit-learn
